@@ -11,7 +11,7 @@ module.exports = {
   },
   dev: {
     env: require('./dev.env'),
-    port: 8080,
+    port: 7788,
     proxyTable: {}
   },
   umd: {
